requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Deep learning
tensorflow>=2.10.0
tensorflow-gpu>=2.10.0  # Optional, for GPU support

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=6.0
python-dateutil>=2.8.0

# Development and testing
pytest>=6.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.991

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Optional: Additional utilities
tqdm>=4.60.0
python-dotenv>=0.19.0